Flames in the Darkness, Deployment Mobile View
Mobile layout for a deployed list.  This still needs more work.  I'd like to redo a the tiles and their information more.  Right now, it's just to get something that works good enough.
Flames in the Darkness list building added. Vengeance gets added to units in the army if a FitD caster is chosen.  Options are displayed to pick 2 Cygnar/Khador warjacks similary to Heart of Darkness.
Selecting a modular Warjack/Warbeast that has only 1 choice per option will have those choices auto-picked.
Master Infernalist added to Journeyman checks and treated as a jack marshal with a limit to one cohort.

Various bug fixes as well.