Deployed List layout added, widgets updated
Deployed Lists additional layout added for mid-size resolutions.
Deployed cohorts now display their systems under their HP.  Warbeasts and Horrors systems are removed when they are disabled.  Warjack systems turn red when they are disabled.
Added a horizontal scrollbar to the HP area of a model's stats for cases like the Colossal warjack damage grids.
Added a Deploy button to the Edit List screen.  It is disabled while the list has not been saved.  You need to save the list first before it turns green to deploy the list.
Resized Horror's web.

Found and fixed an issue where a list would save regardless if there was an error with the list. Current error checks are only if the list has at least 1 leader and if the list doesn't have a name.